+++ src/replacer.js
@@ -5,7 +5,7 @@ const normalizeWeights = _rand.normalizeWeights;
 const weightedChoose = _rand.weightedChoose;
 
 class Replacer {
-  constructor(choices, includeNoOp, identifier) {
+  constructor(choices, includeNoOp, identifier, isSilent) {
     if (includeNoOp) {
       let choicesWithNoOp = choices.slice();
       choicesWithNoOp.push(new WeightedChoice(noOp, null));
@@ -14,6 +14,7 @@ class Replacer {
       this.weights = normalizeWeights(choices);
     }
     this.identifier = identifier;
+    this.isSilent = isSilent;
   }
 
   /**